requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
astroid==3.0.1
bson==0.5.10
click==8.1.7
dill==0.3.7
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
drf-yasg==1.21.7
environ==1.0
fastapi==0.103.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
inflection==0.5.1
isort==5.12.0
mccabe==0.7.0
packaging==23.2
platformdirs==4.0.0
psycopg==3.1.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pylint==3.0.2
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tomlkit==0.12.3
typing_extensions==4.8.0
uritemplate==4.1.1
uvicorn==0.23.2
